requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
gunicorn==25.1.0
idna==3.11
packaging==26.0
pycryptodome==3.23.0
python-dotenv==1.2.1
rave-python==1.4.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.4.0
markdown
Pillow
django-storages[boto3]
boto3
django-markdownx
